+++ bin/Game.cpp
@@ -87,7 +87,7 @@ void Game::updateGame(float deltaS,sf::RenderWindow &game){
 }
 
 void Game::registerDelegates() {
-  //bind our delegate function for key presses
+  //bind our delegate function for speed changes
   EventManager::EventDelegate speedChangeDelegate= std::bind(&Game::handleSpeedChange, this, _1);
 
   //make an event and get its type
@@ -95,6 +95,34 @@ void Game::registerDelegates() {
   EventType speedChangeEventType = speedChangeEvent.getEventType();
   //register the delegate and its type
   this -> eventManager -> registerDelegate(speedChangeDelegate, textLoader -> getString(string("IDS_Game_SpeedChange")),speedChangeEventType);
+
+  //bind our delegate function map generation
+  EventManager::EventDelegate mapGeneratedEventDelegate = std::bind(&Game::handleMapGeneration, this, _1);
+
+  //make an event and get its type
+  MapGeneratedEvent mapGeneratedEvent = MapGeneratedEvent();
+  EventType mapGeneratedEventType = mapGeneratedEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(mapGeneratedEventDelegate, textLoader -> getString(string("IDS_Game_Map_Generated_Delegate")),mapGeneratedEventType);
+
+  //bind our delegate function for level changes
+  EventManager::EventDelegate levelChangeEventDelegate= std::bind(&Game::handleLevelChange, this, _1);
+
+  //make an event and get its type
+  LevelChangeEvent levelChangeEvent = LevelChangeEvent();
+  EventType levelChangeEventType = levelChangeEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(levelChangeEventDelegate, textLoader -> getString(string("IDS_Game_Level_Change_Delegate")),levelChangeEventType);
+
+  //bind our delegate function for game restarts
+  EventManager::EventDelegate restartGameEventDelegate= std::bind(&Game::handleRestartGame, this, _1);
+
+  //make an event and get its type
+  RestartGameEvent restartGameEvent = RestartGameEvent();
+  EventType restartGameEventType = restartGameEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(restartGameEventDelegate, textLoader -> getString(string("IDS_Game_Restarted_Delegate")),restartGameEventType);
+
 }
 
 void Game::deregisterDelegates() {
@@ -111,4 +139,23 @@ void Game::handleSpeedChange(const EventInterface &event) {
   SpeedChangeEventData* scEventData = static_cast<SpeedChangeEventData*>((scEvent -> data).get());
 
   speedScale=scEventData->newSpeedScale;
+}
+
+void Game::handleMapGeneration(const EventInterface &event){
+
+}
+
+/*
+ * Handle Levels (avoids race conditions)
+ */
+void Game::handleLevelChange(const EventInterface &event){
+  gameLogic->handleLevelChangeEvent(event);
+  userView->handleLevelChangeEvent(event);
+}
+
+/*
+ * Handle Restarts (avoids race conditions)
+ */
+void Game::handleRestartGame(const EventInterface &event){
+  gameLogic->handleRestartGameEvent(event);
 }