+++ include/PlayingScreen.hpp
@@ -121,4 +121,6 @@ public:
   shared_ptr<PlayingScreenHeader> getHeader(){return playingScreenHeader;};
 
   bool isFlagButtonClicked(){return rallyPointChange->isButtonClicked();}
+
+  void handleLevelChangeEvent(const EventInterface& event);
 };