+++ Changelog.md
@@ -1,5 +1,9 @@
 [Unreleased]
 
+###Added
+- loading screen with hints[@andrewkatson][https://github.com/andrewkatson]
+
+
 ###Changed
 - buy and sell buttons will mirror across middle of screen if you right click them (useful when they block clicks) [@andrewkatson][https://github.com/andrewkatson]